insert method added to dsr

+++ Natural_Core/Models/Dsr.cs
@@ -7,11 +7,11 @@ namespace Natural_Core.Models
 {
     public partial class Dsr
     {
+
         public Dsr()
         {
             Dsrdetails = new HashSet<Dsrdetail>();
         }
-
         public string Id { get; set; }
         public string Executive { get; set; }
         public string Distributor { get; set; }
@@ -19,13 +19,12 @@ namespace Natural_Core.Models
         public string OrderBy { get; set; }
         public DateTime CreatedDate { get; set; }
         public DateTime ModifiedDate { get; set; }
-        public double TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
+
         public virtual Distributor DistributorNavigation { get; set; }
         public virtual Executive ExecutiveNavigation { get; set; }
         public virtual Login OrderByNavigation { get; set; }
         public virtual Retailor RetailorNavigation { get; set; }
         public virtual ICollection<Dsrdetail> Dsrdetails { get; set; }
-        //public virtual IEnumerable<Product> ProductDetails { get; set; }
-
     }
 }
